Delete __pycache__/ after unit test is done

+++ test/test_application.py
@@ -13,6 +13,7 @@ class TestApplication(unittest.TestCase):
         self.dirname     = os.path.join('.', 'test', 'tmp')
         if not os.path.isdir(self.dirname):
             os.makedirs(self.dirname)
+        self.pycaches = glob.glob(os.path.join('.', '**', '__pycache__'))
 
     def test_webdriver(self):
         if type(self) is TestApplication:
@@ -24,9 +25,15 @@ class TestApplication(unittest.TestCase):
 
     def tearDown(self):
         if type(self) is TestApplication:
-            pass
+            for pycache in self.pycaches:
+                if os.path.isdir(pycache):
+                    shutil.rmtree(pycache)
         else:
-            shutil.rmtree(self.dirname)
+            if os.path.isdir(self.dirname):
+                shutil.rmtree(self.dirname)
+            for pycache in self.pycaches:
+                if os.path.isdir(pycache):
+                    shutil.rmtree(pycache)
 
 if __name__ == '__main__':
     unittest.main()
